+++ vaadin-overlay.d.ts
@@ -1,2 +1,2 @@
 export * from './src/vaadin-overlay.js';
-export * from './@types/interfaces';
+export * from './src/interfaces';
